Add files via upload
Added SHAP and LIME for Model Explainability

In this commit, I enhanced the code to include SHAP and LIME for model explainability. SHAP values were used to provide a unified measure of feature importance, allowing us to understand the contribution of each feature to the model's predictions. Additionally, LIME was utilized to explain individual predictions by approximating the model locally with an interpretable model.

To incorporate SHAP into the code, I installed the 'shap' library and calculated SHAP values for the Decision Tree model. I generated SHAP summary plots, force plots, and dependence plots to visualize the feature importance and relationship between features and model output. These plots were saved as artifacts and logged in MLflow.